Bug 540809 - [GTK4] Replace gdk_test_* functions
- Replaced gdk_test_* functions to simulate mouse/key events with
manually constructed gdkEvents in Display.post().
- Get the currently focused gdkWindow for keyboard events with
gdk_get_default_root_window and gdk_window_get_children instead of using
GdkScreen.
- Rewrote logic for accumulating keyboard state and keyval. Rewrote
caching mechanism for modifier states.
- Fixed several type cast discrepancies with native methods. Moved
OS.gdk_keymap_translate_keyboard_state to GDK.java

Side effects:
- Key events can be posted on Wayland as well due to new method for
getting focused gdkWindow
- Also partially fixed bug 521506, button pressed signal can be emitted
(but not menu items). Maybe also made bug 540441 better.
- Known limitation: mouse events cannot be simulated on Wayland because:
1) gdk_device_warp does not have an implementation on wayland, cursor
cannot be moved. 2) gdk_device_get_window_at_position returns null on
Wayland because pointer device is not known to GDK.

Tested with Bug302171_AltGr (x11/Wayland),
Bug537025_SendKeyEvent(x11/Wayland), Bug521506_FakeClick(x11 only)

+++ bundles/org.eclipse.swt/Eclipse SWT PI/gtk/org/eclipse/swt/internal/gtk/GdkKeymapKey.java
@@ -16,7 +16,7 @@ package org.eclipse.swt.internal.gtk;
 
 public class GdkKeymapKey {
 	/** @field cast=(guint) */
-	public long keycode;
+	public int keycode;
 	/** @field cast=(gint) */
 	public int group;
 	/** @field cast=(gint) */
